\name{WeatherMap.aspect}
\alias{WeatherMap.aspect}
\title{Aspect ratio}
\usage{
  WeatherMap.aspect(Options)
}
\arguments{
  \item{Options}{list of options - see
  \code{WeatherMap.set.option}}
}
\value{
  ratio
}
\description{
  Appropriate aspect ratio - for device allocation
}
\details{
  Get an aspect ratio (width/height) for the region set in
  the options.
}

