% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WeatherMap.draw.haze}
\alias{WeatherMap.draw.haze}
\title{Haze to indicate altitude.}
\usage{
WeatherMap.draw.haze(height, Options)
}
\arguments{
\item{height}{height to simulate (hpa)}

\item{Options}{list of options - see \code{WeatherMap.set.option}}
}
\value{
nothing - side effect only.
}
\description{
Draw the haze - semi transparent blue
}
\details{
As it's a semi-transparent area field we need to draw in
 pixel coordinates to rule out gaps or overlaps, so
  use an image.
}

