% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WeatherMap.R
\name{WeatherMap.draw.label}
\alias{WeatherMap.draw.label}
\title{Draw label}
\usage{
WeatherMap.draw.label(Options)
}
\arguments{
\item{Options}{list of options - see \code{WeatherMap.set.option}}
}
\value{
nothing - side effect only.
}
\description{
Add a label - usually the date
}
\details{
Position and content set in options.
}

