% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WeatherMap.R
\name{WeatherMap.draw.obs}
\alias{WeatherMap.draw.obs}
\title{Draw observations}
\usage{
WeatherMap.draw.obs(obs, Options)
}
\arguments{
\item{obs}{data frame from \code{GSDF.TWCR::TWCR.get.obs} or equivalent.}

\item{Options}{list of options - see \code{WeatherMap.set.option}}
}
\value{
nothing - side effect only.
}
\description{
Draw a set of observations - as points
}
\details{
Only expected to be used with 20CR observations, so take observations
 in the prepbufr format generated by \code{GSDF.TWCR::TWCR.get.obs}.
 That is a data frame - we're only interested in columns V4 (lon) and
 V5 (lat).
}

