% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WeatherMap.R
\name{WeatherMap.make.streamlines}
\alias{WeatherMap.make.streamlines}
\title{Make streamlines}
\usage{
WeatherMap.make.streamlines(s, u, v, t, t.c, Options)
}
\arguments{
\item{s}{streamlines from previous timestep (or NULL)}

\item{u}{GSDF field of zonal wind (m/s)}

\item{v}{GSDF field of meridional wind (m/s)}

\item{t}{GSDF field of air temperature (K)}

\item{t.c}{GSDF field of air temperature normal (K)}

\item{Options}{list of options - see \code{WeatherMap.set.option}}
}
\value{
streamlines data structure.
}
\description{
A streamline is the locus of a point moving with the wind. We need to
 seed them, propagate them out with the wind, keep track of the local
 temperature (for colour when drawing), and thin them.
}
\details{
We keep streamline data in a list. If there are 'i' streamlines, each of length 'j',
 components are
  \itemize{
     \item{x}{longitudes, array[i,j]}
     \item{y}{latitudes, array[i,j]}
     \item{status}{vector[i]. For a newly created streamline, status=0. It increases
by 1/Options$wind.vector.fade.steps each timestep to a max of 1. When chosen for thinning,
status is set to -1 and increases similarly to 0, when the streamline is deleted. As line width
is proportional to status, this allows streamlines to fade gradually in and out.}
     \item{shape}{vector[i]. See \code{grid::xspline}, currently always 0.}
     \item{t_anom}{Temperature anomaly at origin (controls plot colour). vector[i]}
  }
This function will move the streamlines on one timestep, seeding new ones as necessary,
 propagating, thinning and updating the status of existing streamlines.
}
\seealso{
\code{WeatherMap.draw.streamlines}
}

