% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WeatherMap.set.option}
\alias{WeatherMap.set.option}
\title{WeatherMap.option}
\usage{
WeatherMap.set.option(Options = NULL, option = NULL, value = NULL)
}
\arguments{
\item{Options}{list of options - if NULL, use defaults}

\item{option}{name of option to set}

\item{value}{value to set selected option to}
}
\value{
new list of options
}
\description{
Set or query the options controling the plot.
}
\details{
The rendering of a map is controlled by a large
 number of options contained in a list.
See source for values and defaults.
}

